Start work with tournament

+++ Chess.Infrastructure/Hubs/ChessMatchHub.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using Chess.Core.Domain;
+using Chess.Core.Domain.Enum;
 using Chess.Infrastructure.Services;
 using Chess.Infrastructure.Settings;
 using Microsoft.AspNetCore.SignalR;
@@ -15,17 +16,23 @@ namespace Chess.Infrastructure.Hubs
         private readonly ChessGameSettings _chessGameSettings;
         private readonly IFileProvider _fileProvider;
         private readonly IUserService _userService;
+        private readonly IELOProvider _eloProvider;
+        private readonly IPlayerService _playerService;
 
         public ChessMatchHub(IChessGameService chessGameService,
                             IOptions<ChessGameSettings> chessGameSettings,
                             IUserService userService,
-                            IFileProvider fileProvider
+                            IFileProvider fileProvider,
+                            IELOProvider eLOProvider,
+                            IPlayerService playerService
                             )
         {
             _chessGameService = chessGameService;
             _chessGameSettings = chessGameSettings.Value;
             _userService = userService;
             _fileProvider = fileProvider;
+            _eloProvider = eLOProvider;
+            _playerService = playerService;
         }
 
         public override async Task OnConnectedAsync() 
@@ -37,24 +44,26 @@ namespace Chess.Infrastructure.Hubs
 
         public override async Task OnDisconnectedAsync(Exception exception)
         {
-            //If player was in waiting list : Remove from waiting list
-            await _chessGameService.RemoveCurrentPlayerFromWaitingList(this.Context.ConnectionId);
-            //If player was in play: Call opponent about disconnect 
-            var roomId = await _chessGameService.GetRoomId(this.Context.ConnectionId);
-            await Clients.Group(roomId).SendAsync("ReceiveCommunication", "Opponent left the game");
 
-            var chessMatch = await _chessGameService.GetChessMatch(roomId);
-            
-            if(!chessMatch.isSaved)
-            {
-                //create pgn file 
-                //save game to database 
-                var fileName = GenerateFileName("aa", "bb","pgn");
-                // chessMatch.Match.PgnPath = fileName;
-                chessMatch.Match.PgnFileName = fileName;
-                await _fileProvider.SaveFile(_chessGameSettings.PGNFilePath, fileName, chessMatch.PGN);
-                await _chessGameService.SaveChessMatchOnDatabase(roomId);
-                chessMatch.isSaved = true;
+            if (await _chessGameService.GetPlayerFromWaitingList(this.Context.ConnectionId) != null){
+                //If player was in waiting list : Remove from waiting list
+                await _chessGameService.RemoveCurrentPlayerFromWaitingList(this.Context.ConnectionId);
+            }
+            else{
+                //If player was in play: Call opponent about disconnect 
+                var roomId = await _chessGameService.GetRoomId(this.Context.ConnectionId);
+                await Clients.Group(roomId).SendAsync("ReceiveCommunication", "Opponent left the game");
+                var chessMatch = await _chessGameService.GetChessMatch(roomId);
+                if(!chessMatch.isSaved)
+                {
+                    //create pgn file 
+                    var fileName = GenerateFileName("aa", "bb","pgn");
+                    // chessMatch.Match.PgnPath = fileName;
+                    chessMatch.Match.PgnFileName = fileName;
+                    await _fileProvider.SaveFile(_chessGameSettings.PGNFilePath, fileName, chessMatch.PGN);
+                    await _chessGameService.SaveChessMatchOnDatabase(roomId);
+                    chessMatch.isSaved = true;
+                }
             }
 
             Console.WriteLine($"[{this.Context.ConnectionId}] is closed");
@@ -88,48 +97,86 @@ namespace Chess.Infrastructure.Hubs
             return this.Context.ConnectionId;
         }
 
-        public async Task<string> SearchRandomChessGame(string gameDuration){
-            int gameDurationParsered = int.Parse(gameDuration);
-            Console.WriteLine(gameDurationParsered); 
-            string userName = Context.User.Identity.Name;
-            Console.WriteLine("--> Connection id: " + this.Context.ConnectionId + " and userId " + userName + " serching game");
-
-            var currentPlayer = new PlayerInRoom(new Guid("c843d008-8610-4959-9b80-9569ff02ddc0"), userName, gameDurationParsered, this.Context.ConnectionId);
-            if(_chessGameService.CountOpponent(gameDurationParsered) == 0)
-            {
-                await _chessGameService.AddToWaitingList(currentPlayer);
-                Console.WriteLine("--> Connected user witch connectionId " + userName + " was added to waiting player list");
-                return "Added to waiting list";
-            }
-            else
-            {
-                //NA tym etapie, powinno być dołczenie do grup i zwrotka do obu graczy ze się zaczeło
-                Console.WriteLine("--> Searching good opponent");
-                var opponent = await _chessGameService.GetPlayerFromWaitingList(gameDurationParsered);
-                Console.WriteLine("--> Connected user witch userId " + userName + " found player with id " + opponent.UserId + " " + opponent.ConnectionId);
-                var room = Room.CreateRoom();
-                Console.WriteLine("--> Create new room with id " + room.Id.ToString());
-
-                Task t1 = Groups.AddToGroupAsync(this.Context.ConnectionId, room.Id.ToString());
-                Task t2 = Groups.AddToGroupAsync(opponent.ConnectionId.ToString(), room.Id.ToString());
-
-                await _chessGameService.InitMatch(currentPlayer, opponent, room.Id.ToString());
-
-                t1.Wait();
-                t2.Wait();
-                
-                await _chessGameService.RemoveFromWaitingList(opponent);
-
-                await Clients.Client(this.Context.ConnectionId).SendAsync("GetColorPiece",ChessboardPieceColor.white.ToString());
-                await  Clients.Client(opponent.ConnectionId).SendAsync("GetColorPiece",ChessboardPieceColor.black.ToString());
-                                
-                await Clients.Client(this.Context.ConnectionId).SendAsync("GetOpponent",opponent.Username);
-                await  Clients.Client(opponent.ConnectionId).SendAsync("GetOpponent",userName);
-                                
-                await Clients.Group(room.Id.ToString()).SendAsync("ReceiveRoom", room.Id.ToString());
-                return "Player found game";
-            }
-        }
+public async Task<string> SearchRandomChessGame(int gameDuration, 
+                                                string playerId){
+    float RaW = 0, RaD = 0, RaL = 0, RbW = 0, RbD = 0, RbL = 0;
+    var currentPlayer = await _playerService
+                        .GetAsync(Context.User.Identity.Name);
+    var currentPlayerInRoom = new PlayerInRoom(currentPlayer, 
+                                             gameDuration,
+                                             this.Context.ConnectionId, 
+                                             SearchingGameTypEnum.random);
+    Console.WriteLine("--> Player " + currentPlayer.Username +
+                    "was added to room");
+    if(_chessGameService.CountOpponent(gameDuration, 
+                                    currentPlayer.RatingElo, 
+                                    SearchingGameTypEnum.random) == 0){
+        await _chessGameService.AddToWaitingList(currentPlayerInRoom);
+        Console.WriteLine("--> " + currentPlayer.Username +
+                        "was added to waiting list");
+        return "Added to waiting list";
+    }
+    else{
+        var opponent = await _chessGameService
+            .GetPlayerFromWaitingList(gameDuration, 
+                                    currentPlayer.RatingElo, 
+                                    SearchingGameTypEnum.random);
+        Console.WriteLine("--> " + currentPlayer.Username + 
+                        "found opponent " + opponent.Username);
+        await _chessGameService
+            .RemoveFromWaitingList(opponent);
+        Console.WriteLine("--> Player " + opponent.Username +
+                        " was removed from waiting list"); 
+        var room = Room.CreateRoom();
+        Console.WriteLine("--> Create new room with id " + 
+                          room.Id.ToString());
+        Task t1 = Groups.AddToGroupAsync(this.Context.ConnectionId, 
+                                        room.Id.ToString());
+        Task t2= Groups.AddToGroupAsync(opponent.ConnectionId.ToString(), 
+                                        room.Id.ToString());
+        await _chessGameService.InitMatch(currentPlayerInRoom, opponent, 
+                                        room.Id.ToString());
+        t1.Wait();
+        t2.Wait();
+        _eloProvider.CalcELORating(currentPlayer.RatingElo, 
+                            opponent.RatingElo,ChessGameResultEnum.WIN, 
+                            ref RaW, ref RbL);
+        _eloProvider.CalcELORating(currentPlayer.RatingElo, 
+                            opponent.RatingElo,ChessGameResultEnum.DRAW, 
+                            ref RaD, ref RbD);
+        _eloProvider.CalcELORating(currentPlayer.RatingElo, 
+                            opponent.RatingElo,ChessGameResultEnum.LOSE, 
+                            ref RaL, ref RbW);
+        await Clients.Client(this.Context.ConnectionId)
+                            .SendAsync("GetNewRatingELO",RaW,RaD,RaL);
+        await Clients.Client(opponent.ConnectionId)
+                            .SendAsync("GetNewRatingELO",RbW,RbD,RbL);
+        var currentPieceColor = await _chessGameService
+                                    .GetRandomColorPiece();
+        var opponentPieceColor = currentPieceColor
+                                .Equals(ChessboardPieceColorEnum.white.ToString()) ?
+                                ChessboardPieceColorEnum.black.ToString() :
+                                ChessboardPieceColorEnum.white.ToString(); 
+        await Clients.Client(this.Context.ConnectionId)
+                            .SendAsync("GetColorPiece", 
+                                    currentPieceColor);
+        await  Clients.Client(opponent.ConnectionId)
+                      .SendAsync("GetColorPiece", 
+                      opponentPieceColor);
+        await Clients.Client(this.Context.ConnectionId)
+                     .SendAsync("GetOpponentInformation", 
+                        opponent.Username, 
+                        opponent.RatingElo);
+        await  Clients.Client(opponent.ConnectionId)
+                      .SendAsync("GetOpponentInformation", 
+                                currentPlayer.Username, 
+                                currentPlayer.RatingElo);
+        await Clients.Group(room.Id.ToString())
+                     .SendAsync("ReceiveRoom",
+                                room.Id.ToString());
+        return "Player found game";
+    }
+}
 
 
         //color: "w"
@@ -154,8 +201,6 @@ namespace Chess.Infrastructure.Hubs
             public string San { get; set; }
         }
 
-        private enum ChessboardPieceColor{
-            white,black
-        }
+
     }
 }
